requirements.txt
google-cloud-bigquery
psycopg2-binary
tenacity
jupyter
lazy-property
